Refactor Program routing into endpoint groups (#34)
* refactor: extract core endpoint mapping

* refactor: extract provider endpoint mapping

* refactor: extract admin endpoint mapping

* Refactor VCS endpoint mapping (#35)

* refactor: extract vcs endpoint mapping

* refactor: extract module endpoint mapping (#36)


---------


---------
