+++ TerraformRegistry/Program.cs
@@ -109,504 +109,11 @@ var port = builder.Configuration["PORT"] ?? builder.Configuration["Port"] ?? "51
 if (!int.TryParse(port, out var portNumber))
     throw new InvalidOperationException($"Invalid port specified: '{port}'. Please check your configuration.");
 
-app.MapGet("/", async context =>
-{
-    var indexPath = Path.Combine(webFolderPath, "index.html");
-    if (File.Exists(indexPath))
-    {
-        context.Response.ContentType = "text/html";
-        await context.Response.SendFileAsync(indexPath);
-    }
-    else
-    {
-        context.Response.StatusCode = 404;
-    }
-});
-
-// OIDC Authentication endpoints
-var oauthService = app.Services.GetRequiredService<OAuthService>();
-
-app.MapGet("/api/auth/providers", () => AuthHandlers.GetProviders(oauthService))
-    .WithTags("Authentication")
-    .WithDescription("Returns list of enabled OIDC providers");
-
-app.MapGet("/api/auth/login/{provider}", (string provider, string? returnTo, HttpContext context) =>
-        AuthHandlers.Login(provider, returnTo, oauthService, context))
-    .WithTags("Authentication")
-    .WithDescription("Initiates OIDC login flow for the specified provider");
-
-app.MapGet("/api/auth/callback/{provider}", async (string provider, string? code, string? state, string? error,
-            HttpContext context, IApiKeyService apiKeyService, IAuditService auditService, ILogger<Program> authLogger) =>
-        await AuthHandlers.Callback(provider, code, state, error, oauthService, jwtService, apiKeyService, auditService, context,
-            authLogger))
-    .WithTags("Authentication")
-    .WithDescription("Handles OIDC callback after provider authentication");
-
-app.MapGet("/api/auth/me", async (HttpContext context, IPermissionService permService) =>
-    {
-        return await AuthHandlers.GetCurrentUser(jwtService, permService, context);
-    })
-    .WithTags("Authentication")
-    .WithDescription("Returns current user info from session");
-
-app.MapPost("/api/auth/logout", (HttpContext context, IAuditService auditService) => AuthHandlers.Logout(context, auditService))
-    .WithTags("Authentication")
-    .WithDescription("Logs out the current user");
-
-app.MapDelete("/api/auth/me", (HttpContext context, IApiKeyService apiKeyService, IDatabaseService dbService, IAuditService auditService) =>
-        AuthHandlers.DeleteAccount(context, apiKeyService, dbService, auditService))
-    .WithTags("Authentication")
-    .WithDescription("Deletes the current user account")
-    .RequireAuthorization();
-
-app.MapGet("/api/auth/session", (HttpContext context) => AuthHandlers.CheckSession(jwtService, context))
-    .WithTags("Authentication")
-    .WithDescription("Checks if user has a valid session");
-
-app.MapGet("/api/auth/terraform/authorize", (HttpContext context) =>
-        AuthHandlers.BeginTerraformAuthorization(context))
-    .WithTags("Authentication")
-    .WithDescription("Begins Terraform CLI OAuth authorization flow");
-
-app.MapPost("/api/auth/terraform/token", (HttpContext context, ITerraformAuthorizationCodeStore codeStore, IApiKeyService apiKeyService, IAuditService auditService) =>
-        AuthHandlers.ExchangeTerraformToken(context, codeStore, apiKeyService, auditService))
-    .WithTags("Authentication")
-    .WithDescription("Exchanges a Terraform CLI OAuth authorization code for an API token");
-
-// Dev-only login endpoint — not registered in production (route doesn't exist)
-if (app.Environment.IsDevelopment())
-{
-    app.MapPost("/api/auth/dev-login",
-            (JwtService jwt, IConfiguration cfg, IHostEnvironment env, IAuditService auditService, HttpContext ctx) =>
-                AuthHandlers.DevLogin(jwt, cfg, env, auditService, ctx))
-        .WithTags("Authentication")
-        .WithDescription("Creates a dev session (Development only, requires TF_REG_DevAuthBypass=true)");
-}
-
-app.MapGet("/.well-known/terraform.json", ServiceDiscoveryHandlers.GetServiceDiscovery)
-    .WithTags("Service Discovery")
-    .WithDescription("Terraform service discovery endpoint")
-    .Produces<ServiceDiscovery>();
-
-app.MapGet("/health", HealthHandlers.HandleHealth)
-    .WithTags("Health")
-    .WithDescription("Liveness probe")
-    .Produces(200);
-
-app.MapGet("/ready", (IDatabaseService dbService, IModuleService moduleService, IProviderArtifactStorage providerArtifactStorage, HttpContext context, IConfiguration config) =>
-        HealthHandlers.HandleReady(dbService, moduleService, providerArtifactStorage, context, config))
-    .WithTags("Health")
-    .WithDescription("Readiness probe — use ?detail=true with auth for component details")
-    .Produces(200)
-    .Produces(503);
-
-// Analytics endpoints (auth handled by middleware via /api/analytics prefix)
-app.MapGet("/api/analytics/downloads/summary", (IAnalyticsService analytics, HttpContext context) =>
-        AnalyticsHandlers.GetSummary(analytics, context))
-    .WithTags("Analytics")
-    .WithDescription("Download summary statistics");
-
-app.MapGet("/api/analytics/downloads/top", (IAnalyticsService analytics, HttpContext context, int limit = 10, string period = "30d") =>
-        AnalyticsHandlers.GetTopModules(analytics, context, limit, period))
-    .WithTags("Analytics")
-    .WithDescription("Top downloaded modules");
-
-app.MapGet("/api/analytics/downloads/trends", (IAnalyticsService analytics, HttpContext context, string period = "30d", string interval = "day") =>
-        AnalyticsHandlers.GetTrends(analytics, context, period, interval))
-    .WithTags("Analytics")
-    .WithDescription("Download trends over time");
-
-app.MapGet("/api/analytics/downloads/module/{namespace}/{name}/{provider}",
-        (string @namespace, string name, string provider, IAnalyticsService analytics, HttpContext context, string period = "30d") =>
-            AnalyticsHandlers.GetModuleAnalytics(@namespace, name, provider, analytics, context, period))
-    .WithTags("Analytics")
-    .WithDescription("Per-module download analytics");
-
-// Provider management endpoints (auth handled by middleware via /api/providers prefix)
-app.MapGet("/api/providers", (IProviderRegistryService service, HttpContext context, string? q, int offset = 0, int limit = 20) =>
-        ProviderHandlers.ListProviders(service, context, q, offset, limit))
-    .WithTags("Providers");
-
-app.MapPost("/api/providers", (IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.CreateProvider(service, context, request))
-    .WithTags("Providers");
-
-app.MapGet("/api/providers/{namespace}/{type}", (string @namespace, string type, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.GetProvider(@namespace, type, service, context))
-    .WithTags("Providers");
-
-app.MapPatch("/api/providers/{namespace}/{type}", (string @namespace, string type, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.UpdateProvider(@namespace, type, service, context, request))
-    .WithTags("Providers");
-
-app.MapDelete("/api/providers/{namespace}/{type}", (string @namespace, string type, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.DeleteProvider(@namespace, type, service, context))
-    .WithTags("Providers");
-
-app.MapGet("/api/providers/{namespace}/{type}/gpg-keys", (string @namespace, string type, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.ListGpgKeys(@namespace, type, service, context))
-    .WithTags("Providers");
-
-app.MapPost("/api/providers/{namespace}/{type}/gpg-keys", (string @namespace, string type, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.AddGpgKey(@namespace, type, service, context, request))
-    .WithTags("Providers");
-
-app.MapDelete("/api/providers/{namespace}/{type}/gpg-keys/{keyId}", (string @namespace, string type, string keyId, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.RevokeGpgKey(@namespace, type, keyId, service, context))
-    .WithTags("Providers");
-
-app.MapGet("/api/providers/{namespace}/{type}/versions", (string @namespace, string type, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.ListVersions(@namespace, type, service, context))
-    .WithTags("Providers");
-
-app.MapPost("/api/providers/{namespace}/{type}/versions", (string @namespace, string type, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.CreateVersion(@namespace, type, service, context, request))
-    .WithTags("Providers");
-
-app.MapDelete("/api/providers/{namespace}/{type}/versions/{version}", (string @namespace, string type, string version, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.DeleteVersion(@namespace, type, version, service, context))
-    .WithTags("Providers");
-
-app.MapGet("/api/providers/{namespace}/{type}/versions/{version}/platforms", (string @namespace, string type, string version, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.ListPlatforms(@namespace, type, version, service, context))
-    .WithTags("Providers");
-
-app.MapPost("/api/providers/{namespace}/{type}/versions/{version}/platforms", (string @namespace, string type, string version, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.CreatePlatform(@namespace, type, version, service, context, request))
-    .WithTags("Providers");
-
-app.MapDelete("/api/providers/{namespace}/{type}/versions/{version}/platforms/{os}/{arch}", (string @namespace, string type, string version, string os, string arch, IProviderRegistryService service, HttpContext context) =>
-        ProviderHandlers.DeletePlatform(@namespace, type, version, os, arch, service, context))
-    .WithTags("Providers");
-
-app.MapPut("/api/providers/{namespace}/{type}/versions/{version}/shasums", (string @namespace, string type, string version, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.UploadShasums(@namespace, type, version, service, context, request))
-    .WithTags("Providers");
-
-app.MapPut("/api/providers/{namespace}/{type}/versions/{version}/shasums.sig", (string @namespace, string type, string version, IProviderRegistryService service, HttpContext context, HttpRequest request) =>
-        ProviderHandlers.UploadShasumsSignature(@namespace, type, version, service, context, request))
-    .WithTags("Providers");
-
-app.MapPut("/api/providers/{namespace}/{type}/versions/{version}/platforms/{os}/{arch}/package",
-        (string @namespace, string type, string version, string os, string arch, IProviderRegistryService service, HttpContext context,
-                HttpRequest request) =>
-            ProviderHandlers.UploadPlatformPackage(@namespace, type, version, os, arch, service, context, request))
-    .WithTags("Providers");
-
-// Webhook endpoints (admin-only, auth handled by middleware via /api/admin prefix)
-app.MapGet("/api/admin/webhooks", (IWebhookService webhookService, HttpContext context) =>
-        WebhookHandlers.ListWebhooks(webhookService, context))
-    .WithTags("Webhooks");
-
-app.MapPost("/api/admin/webhooks", (IWebhookService webhookService, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        WebhookHandlers.CreateWebhook(webhookService, auditService, context, request))
-    .WithTags("Webhooks");
-
-app.MapPut("/api/admin/webhooks/{id}", (Guid id, IWebhookService webhookService, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        WebhookHandlers.UpdateWebhook(id, webhookService, auditService, context, request))
-    .WithTags("Webhooks");
-
-app.MapDelete("/api/admin/webhooks/{id}", (Guid id, IWebhookService webhookService, IAuditService auditService, HttpContext context) =>
-        WebhookHandlers.DeleteWebhook(id, webhookService, auditService, context))
-    .WithTags("Webhooks");
-
-app.MapPost("/api/admin/webhooks/{id}/test", (Guid id, IWebhookService webhookService, WebhookDispatcher dispatcher, HttpContext context) =>
-        WebhookHandlers.TestWebhook(id, webhookService, dispatcher, context))
-    .WithTags("Webhooks");
-
-// Admin - Roles
-app.MapGet("/api/admin/roles", (IRoleService roleService, HttpContext context) => AdminHandlers.ListRoles(roleService, context)).WithTags("Admin");
-app.MapPost("/api/admin/roles", (IRoleService roleService, IAuditService auditService, HttpContext context, HttpRequest request) => AdminHandlers.CreateRole(roleService, auditService, context, request)).WithTags("Admin");
-app.MapPut("/api/admin/roles/{id}", (Guid id, IRoleService roleService, IAuditService auditService, HttpContext context, HttpRequest request) => AdminHandlers.UpdateRole(id, roleService, auditService, context, request)).WithTags("Admin");
-app.MapDelete("/api/admin/roles/{id}", (Guid id, IRoleService roleService, IAuditService auditService, HttpContext context) => AdminHandlers.DeleteRole(id, roleService, auditService, context)).WithTags("Admin");
-
-// Admin - Audit
-app.MapGet("/api/admin/audit", (IAuditService auditService, HttpContext context, string? action, string? userId, string? resourceType, DateTime? from, DateTime? to, int limit = 50, int offset = 0) =>
-    AuditHandlers.ListAuditLogs(auditService, context, action, userId, resourceType, from, to, limit, offset)).WithTags("Admin");
-app.MapGet("/api/admin/audit/{id}", (Guid id, IAuditService auditService, HttpContext context) =>
-    AuditHandlers.GetAuditLog(id, auditService, context)).WithTags("Admin");
-
-// Admin - Users
-app.MapGet("/api/admin/users", (IDatabaseService dbService, IPermissionService permService, HttpContext context) => AdminHandlers.ListUsers(dbService, permService, context)).WithTags("Admin");
-app.MapGet("/api/admin/users/{userId}/roles", (string userId, IPermissionService permService, HttpContext context) => AdminHandlers.GetUserRoles(userId, permService, context)).WithTags("Admin");
-app.MapPost("/api/admin/users/{userId}/roles", (string userId, IPermissionService permService, IAuditService auditService, HttpContext context, HttpRequest request) => AdminHandlers.AssignUserRole(userId, permService, auditService, context, request)).WithTags("Admin");
-app.MapDelete("/api/admin/users/{userId}/roles/{roleId}", (string userId, Guid roleId, IPermissionService permService, IRoleService roleService, IAuditService auditService, HttpContext context) => AdminHandlers.RemoveUserRole(userId, roleId, permService, roleService, auditService, context)).WithTags("Admin");
-
-// Admin - Module Docs
-app.MapGet("/api/admin/module-docs/summary",
-        (IDatabaseService dbService, IModuleExtractionConfigService configService, HttpContext context) =>
-            ModuleDocsHandlers.GetSummary(dbService, configService, context))
-    .WithTags("Module Docs");
-app.MapGet("/api/admin/module-docs/modules",
-        (IDatabaseService dbService, HttpContext context, string? status, string? q, int limit = 50, int offset = 0) =>
-            ModuleDocsHandlers.ListModules(dbService, context, status, q, limit, offset))
-    .WithTags("Module Docs");
-app.MapGet("/api/admin/module-docs/modules/{namespace}/{name}/{provider}/{version}",
-        (string @namespace, string name, string provider, string version, IDatabaseService dbService, HttpContext context) =>
-            ModuleDocsHandlers.GetModuleDetail(@namespace, name, provider, version, dbService, context))
-    .WithTags("Module Docs");
-app.MapPost("/api/admin/module-docs/modules/{namespace}/{name}/{provider}/{version}/regenerate-llm",
-        (string @namespace, string name, string provider, string version, IModuleExtractionService extractionService,
-                IDatabaseService dbService, IAuditService auditService, IModuleExtractionConfigService configService,
-                HttpContext context) =>
-            ModuleDocsHandlers.RegenerateLlmContext(@namespace, name, provider, version, extractionService, dbService,
-                auditService, configService, context))
-    .WithTags("Module Docs");
-app.MapPost("/api/admin/module-docs/modules/{namespace}/{name}/{provider}/{version}/requeue",
-        (string @namespace, string name, string provider, string version, IModuleExtractionService extractionService,
-                IDatabaseService dbService, IAuditService auditService, IModuleExtractionConfigService configService,
-                HttpContext context) =>
-            ModuleDocsHandlers.Requeue(@namespace, name, provider, version, extractionService, dbService, auditService,
-                configService, context))
-    .WithTags("Module Docs");
-app.MapPost("/api/admin/module-docs/backfill",
-        (IModuleExtractionService extractionService, IModuleExtractionConfigService configService,
-                IAuditService auditService, HttpContext context, HttpRequest request) =>
-            ModuleDocsHandlers.Backfill(extractionService, configService, auditService, context, request))
-    .WithTags("Module Docs");
-app.MapGet("/api/admin/module-docs/config",
-        (IModuleExtractionConfigService configService, HttpContext context) =>
-            ModuleDocsHandlers.GetConfig(configService, context))
-    .WithTags("Module Docs");
-app.MapPut("/api/admin/module-docs/config",
-        (IModuleExtractionConfigService configService, IAuditService auditService, HttpContext context,
-                HttpRequest request) =>
-            ModuleDocsHandlers.UpdateConfig(configService, auditService, context, request))
-    .WithTags("Module Docs");
-
-// VCS source CRUD endpoints (auth handled by middleware via /api/vcs/sources prefix)
-app.MapGet("/api/vcs/sources", (IVcsSourceService vcsService, HttpContext context) =>
-        VcsHandlers.ListVcsSources(vcsService, context))
-    .WithTags("VCS");
-
-app.MapPost("/api/vcs/sources", (IVcsSourceService vcsService, IVcsConnectionService connectionService, IGitHubVcsService githubService, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        VcsHandlers.CreateVcsSource(vcsService, connectionService, githubService, auditService, context, request))
-    .WithTags("VCS");
-
-app.MapGet("/api/vcs/sources/module/{namespace}/{name}/{provider}", (string @namespace, string name, string provider, IVcsSourceService vcsService, HttpContext context) =>
-        VcsHandlers.GetVcsSourceByModule(vcsService, context, @namespace, name, provider))
-    .WithTags("VCS");
-
-app.MapPut("/api/vcs/sources/{id}", (Guid id, IVcsSourceService vcsService, IVcsConnectionService connectionService, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        VcsHandlers.UpdateVcsSource(id, vcsService, connectionService, auditService, context, request))
-    .WithTags("VCS");
-
-app.MapDelete("/api/vcs/sources/{id}", (Guid id, IVcsSourceService vcsService, IAuditService auditService, HttpContext context) =>
-        VcsHandlers.DeleteVcsSource(id, vcsService, auditService, context))
-    .WithTags("VCS");
-
-app.MapPost("/api/vcs/sources/{id}/sync", (Guid id, IVcsSourceService vcsService, IGitHubVcsService githubService, HttpContext context, HttpRequest request) =>
-        VcsHandlers.SyncVcsSource(id, vcsService, githubService, context, request))
-    .WithTags("VCS");
-
-// VCS Connection admin endpoints
-app.MapGet("/api/admin/vcs-connections", (IVcsConnectionService connectionService, HttpContext context) =>
-        VcsHandlers.ListConnections(connectionService, context))
-    .WithTags("VCS");
-
-app.MapPost("/api/admin/vcs-connections", (IVcsConnectionService connectionService, IConfiguration config, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        VcsHandlers.CreateConnection(connectionService, config, auditService, context, request))
-    .WithTags("VCS");
-
-app.MapPut("/api/admin/vcs-connections/{id}", (Guid id, IVcsConnectionService connectionService, IConfiguration config, IAuditService auditService, HttpContext context, HttpRequest request) =>
-        VcsHandlers.UpdateConnection(id, connectionService, config, auditService, context, request))
-    .WithTags("VCS");
-
-app.MapDelete("/api/admin/vcs-connections/{id}", (Guid id, IVcsConnectionService connectionService, IAuditService auditService, HttpContext context) =>
-        VcsHandlers.DeleteConnection(id, connectionService, auditService, context))
-    .WithTags("VCS");
-
-// Lightweight connection list for Add Module dropdown (auth required, not admin-only)
-app.MapGet("/api/vcs/connections", (IVcsConnectionService connectionService, HttpContext context) =>
-        VcsHandlers.ListConnectionSummaries(connectionService, context))
-    .WithTags("VCS");
-
-// GitHub webhook endpoint (public, no auth required)
-app.MapPost("/api/vcs/github/webhook", (IGitHubVcsService githubService, HttpContext context) =>
-        VcsHandlers.HandleGitHubWebhook(githubService, context))
-    .WithTags("VCS");
-
-app.MapGet("/v1/modules",
-        (IModuleService moduleService, HttpContext context, string? q, string? @namespace, string? provider, int offset = 0,
-                int limit = 10) =>
-            ModuleHandlers.ListModules(moduleService, context, q, @namespace, provider, offset, limit))
-    .WithTags("Modules")
-    .WithDescription("Lists or searches modules")
-    .Produces<ModuleList>();
-
-app.MapGet("/v1/modules/{namespace}/{name}/{provider}/{version}", (string @namespace, string name, string provider,
-            string version, IModuleService moduleService, HttpContext context) =>
-        ModuleHandlers.GetModule(@namespace, name, provider, version, moduleService, context))
-    .WithTags("Modules")
-    .WithDescription("Gets a specific module")
-    .Produces<Module>()
-    .ProducesProblem(404);
-
-app.MapGet("/v1/modules/{namespace}/{name}/{provider}/versions",
-        (string @namespace, string name, string provider, IModuleService moduleService, HttpContext context) =>
-            ModuleHandlers.GetModuleVersions(@namespace, name, provider, moduleService, context))
-    .WithTags("Modules")
-    .WithDescription("Gets all versions of a specific module")
-    .Produces<ModuleVersions>();
-
-app.MapGet("/llm.txt", (IConfiguration configuration) => LlmHandlers.GetGuide(configuration))
-    .WithTags("LLM")
-    .WithDescription("Provides LLM discovery guidance");
-
-app.MapGet("/v1/llm/modules",
-        (IModuleService moduleService, IConfiguration configuration, HttpContext context, string? q, int offset = 0,
-                int limit = 50) =>
-            LlmHandlers.ListModules(moduleService, configuration, context, q, offset, limit))
-    .WithTags("LLM")
-    .WithDescription("Lists modules for authenticated LLM discovery")
-    .Produces<ModuleLlmIndexResponse>();
-
-app.MapGet("/v1/llm/modules/{namespace}/{name}/{provider}",
-        (string @namespace, string name, string provider, IModuleService moduleService, IDatabaseService dbService,
-                IConfiguration configuration, HttpContext context) =>
-            LlmHandlers.GetModuleVersions(@namespace, name, provider, moduleService, dbService, configuration, context))
-    .WithTags("LLM")
-    .WithDescription("Lists module versions and LLM readiness for authenticated clients")
-    .Produces<ModuleLlmModuleVersionsResponse>();
-
-app.MapGet("/v1/llm/modules/{namespace}/{name}/{provider}/{version}",
-        (string @namespace, string name, string provider, string version, IDatabaseService dbService,
-                HttpContext context) =>
-            LlmHandlers.GetModuleContext(@namespace, name, provider, version, dbService, context))
-    .WithTags("LLM")
-    .WithDescription("Returns the stored LLM context for a module version")
-    .Produces<ModuleLlmContextDocument>()
-    .ProducesProblem(409);
-
-app.MapGet("/v1/modules/{namespace}/{name}/{provider}/{version}/download", (string @namespace, string name,
-            string provider, string version, IModuleService moduleService, IDatabaseService dbService, HttpContext context) =>
-        ModuleHandlers.DownloadModule(@namespace, name, provider, version, moduleService, dbService, context))
-    .WithTags("Modules")
-    .WithDescription("Downloads a specific module version")
-    .Produces(200, contentType: "application/zip")
-    .ProducesProblem(404);
-
-app.MapGet("/v1/modules/{namespace}/{name}/{provider}/download",
-        (string @namespace, string name, string provider, IModuleService moduleService, IDatabaseService dbService, HttpContext context) =>
-            ModuleHandlers.DownloadLatestModule(@namespace, name, provider, moduleService, dbService, context))
-    .WithTags("Modules")
-    .WithDescription("Downloads the latest version of a module for a provider")
-    .Produces(302)
-    .ProducesProblem(404);
-
-app.MapGet("/v1/providers/{namespace}/{type}/versions",
-        (string @namespace, string type, IProviderRegistryService providerService, HttpContext context) =>
-            ProviderHandlers.GetVersions(@namespace, type, providerService, context))
-    .WithTags("Providers")
-    .WithDescription("Gets all versions for a provider")
-    .Produces<ProviderVersionsResponse>()
-    .ProducesProblem(404);
-
-app.MapGet("/v1/providers/{namespace}/{type}/{version}/download/{os}/{arch}",
-        (string @namespace, string type, string version, string os, string arch, IProviderRegistryService providerService,
-                HttpContext context) =>
-            ProviderHandlers.GetPackage(@namespace, type, version, os, arch, providerService, context))
-    .WithTags("Providers")
-    .WithDescription("Gets package metadata for a provider version and platform")
-    .Produces<ProviderPackageResponse>()
-    .ProducesProblem(404);
-
-app.MapPost("/v1/modules/{namespace}/{name}/{provider}/{version}", async (string @namespace, string name,
-            string provider, string version, HttpRequest request, IModulePublishCoordinator publishCoordinator, HttpContext context) =>
-        await ModuleHandlers.UploadModule(@namespace, name, provider, version, request, publishCoordinator, context))
-    .WithTags("Modules")
-    .WithDescription("Uploads a new module version")
-    .Accepts<IFormFile>("multipart/form-data")
-    .ProducesProblem(400)
-    .ProducesProblem(409)
-    .Produces(201);
-
-// Module version management - soft delete, restore, purge
-app.MapDelete("/v1/modules/{namespace}/{name}/{provider}/{version}",
-        (string @namespace, string name, string provider, string version, IModuleService moduleService, WebhookDispatcher webhookDispatcher, IAuditService auditService, HttpContext context) =>
-            ModuleHandlers.DeleteModuleVersion(@namespace, name, provider, version, moduleService, webhookDispatcher, auditService, context))
-    .WithTags("Modules")
-    .WithDescription("Soft deletes a module version")
-    .Produces(204)
-    .ProducesProblem(404);
-
-app.MapPost("/v1/modules/{namespace}/{name}/{provider}/{version}/restore",
-        (string @namespace, string name, string provider, string version, IModuleService moduleService, WebhookDispatcher webhookDispatcher, IAuditService auditService, HttpContext context) =>
-            ModuleHandlers.RestoreModuleVersion(@namespace, name, provider, version, moduleService, webhookDispatcher, auditService, context))
-    .WithTags("Modules")
-    .WithDescription("Restores a soft-deleted module version")
-    .Produces(204)
-    .ProducesProblem(404);
-
-app.MapDelete("/v1/modules/{namespace}/{name}/{provider}/{version}/purge",
-        (string @namespace, string name, string provider, string version, IModuleService moduleService, WebhookDispatcher webhookDispatcher, IAuditService auditService, HttpContext context) =>
-            ModuleHandlers.PurgeModuleVersion(@namespace, name, provider, version, moduleService, webhookDispatcher, auditService, context))
-    .WithTags("Modules")
-    .WithDescription("Permanently deletes a module version")
-    .Produces(204)
-    .ProducesProblem(404);
-
-app.MapGet("/v1/modules/trash",
-        (IModuleService moduleService, HttpContext context, string? q, string? @namespace, string? provider, int offset = 0,
-                int limit = 10) =>
-            ModuleHandlers.ListDeletedModules(moduleService, context, q, @namespace, provider, offset, limit))
-    .WithTags("Modules")
-    .WithDescription("Lists all soft-deleted modules")
-    .Produces<ModuleList>();
-
-app.MapPatch("/v1/modules/{namespace}/{name}/{provider}/description",
-        (string @namespace, string name, string provider, HttpRequest request, IModuleService moduleService, IAuditService auditService, HttpContext context) =>
-            ModuleHandlers.UpdateDescription(@namespace, name, provider, request, moduleService, auditService, context))
-    .WithTags("Modules")
-    .WithDescription("Updates the description for a module")
-    .Produces(200)
-    .ProducesProblem(404);
-
-app.MapGet("/module/download", async context =>
-{
-    var token = context.Request.Query["token"].ToString();
-    if (string.IsNullOrEmpty(token) || !LocalModuleService.TryGetFilePathFromToken(token, out var filePath))
-    {
-        context.Response.StatusCode = 404;
-        await context.Response.WriteAsync("Invalid or expired download link.");
-        return;
-    }
-
-    if (!File.Exists(filePath))
-    {
-        context.Response.StatusCode = 404;
-        await context.Response.WriteAsync("File not found.");
-        return;
-    }
-
-    context.Response.ContentType = "application/zip";
-    context.Response.Headers["Content-Disposition"] = $"attachment; filename=\"{Path.GetFileName(filePath)}\"";
-    await context.Response.SendFileAsync(filePath);
-});
-
-app.MapGet("/provider/download", async context =>
-{
-    var token = context.Request.Query["token"].ToString();
-    if (string.IsNullOrEmpty(token) || !LocalProviderArtifactStorage.TryGetFilePathFromToken(token, out var filePath))
-    {
-        context.Response.StatusCode = 404;
-        await context.Response.WriteAsync("Invalid or expired download link.");
-        return;
-    }
-
-    if (!File.Exists(filePath))
-    {
-        context.Response.StatusCode = 404;
-        await context.Response.WriteAsync("File not found.");
-        return;
-    }
-
-    context.Response.ContentType = Path.GetExtension(filePath).Equals(".zip", StringComparison.OrdinalIgnoreCase)
-        ? "application/zip"
-        : "text/plain";
-    context.Response.Headers["Content-Disposition"] = $"attachment; filename=\"{Path.GetFileName(filePath)}\"";
-    await context.Response.SendFileAsync(filePath);
-});
+app.MapCoreEndpoints(webFolderPath, jwtService);
+app.MapProviderEndpoints();
+app.MapAdminEndpoints();
+app.MapVcsEndpoints();
+app.MapModuleEndpoints();
 
 app.MapControllers();
